ghoul applygameoptions
